[usdrecord] Updated usdrecord for PySide6
Updated usdrecord for PySide6 to use an explicitly created
QOpenGLContext bound to a QOffscreenSurface so that we don't
have to depend on QOpenGLWidget creating and binding a context.

Without this fix, usdrecord can crash on startup while trying
to initialize the frameRecorder.

Tested with both PySide6 (6.2.x) and PySide2 (5.15.x).

(Internal change: 2269000)