Updating the UsdStage change processing for prim type name and API schema changes to no longer do a full recompose of the affected prim and only do an update of the prim's UsdPrimTypeInfo.
The full recompose was a huge overkill especially because it would cause
a recompose of what the prim's children are. This is both expensive and
unnecessary since a type name or API schema change will never change the
prim's children. This change results in close to a 100x speed up of the
ApplyAPI operation on a UsdPrim.

Fixes #2343

(Internal change: 2269211)