This is a proper fix for the double transform issue.
To understand the fix, the main insight is that the flattening scene
index introduced by the UsdImaging_NiInstanceAggregationSceneIndex is
there so that it can read the flattened transforms and then update the
instancer's transform primvar.  The flatteningSceneIndex does not need
to be part of the main scene index chain.

This fix updates the _InstanceObserver to use the flattening scene index
as an input so that it reads flattened transform values on the natively
instanced prims.  These flatteend transform values get authored to the
instancer.  Now, clients provide a normal flattening scene index to
flatten in the scene index chain.

Note, this fix also includes a change to the niPrototypeSceneIndex.  We
need this to make sure the prototypes are always at the identity.
Without this, the flattening scene index may push an "overrides"
transform onto the prototypes themselves.

(Internal change: 2269514)