Update NormalMapTexture compliance checker and UsdUVTexture/UsdPreviewSurface docs update
Instead on Warning, the compliance checker rules Errors when the following rules are violated:
- Shader prims are not encapsulated with in a container type prim.
- For 8 bit normals map:
  - if sourceColorSpace is not "raw"
  - if bias and scale are not provided.

Compliance checker still warns when scale is not (2, 2, 2, 1) for 8 bit normal map, as some authoring
tools use this for normal pertubation. Additionally if the scale values are compliant but If bias is not
(-1, -1, -1, 0), clients are warned.

Updates are also made to shader docs to reflect these requirements which the compliance checker errors on.

(Internal change: 2269650)