Revert "[HdSt, Hdx, HgiMetal] Full Masking Viewport issue on Mac"
The intention of the submit ab0520a2 was to make the ComputeViewport the same in OpenGL and Metal.

Originally, in Metal, ComputeViewport will just output the dataWIndow as the viewport. And in HgiMetalGraphicsCmds::SetViewport, it uses "viewportHeight - viewportMinY" as the final minY. So we can know:
finalMinY = viewportHeight - viewportMinY = windowHeight - windowMinY.

In the submit of ab0520a2, it uses the OpenGL calculation in ComputeViewport for both OpenGL and Metal. So ComputeViewport will use "frameBufferHeight - windowMinY - windowHeight" as the minY of viewport. And in HgiMetalGraphicsCmds::SetViewport, it uses "viewportHeight + viewportMin" as the final minY. So now :
finalMinY = viewportHeight + viewportMinY = windowHeight + frameBufferHeight - windwMinY - windowHeight = framebufferHeight - windowMinY.

Indeed after this change, the finalMinY is different from the original finalMinY. If framebufferHeight is the same as the windowHeight, there is no problem. If they are different, the rendering result will be indirect.

So here we just revert the change.