HdRenderPassState: making the calls to set the camera, override window policy and framing/viewport orthogonal.
Respecting the override window policy whether using framing or viewport.

testUsdImagingGLBasicDrawing explicitly takes a window policy which default to "match vertically" to match what usdview is doing.

This is motivated by several tests failing when enabling USDIMAGINGGL_ENGINE_ENABLE_SCENE_INDEX because of resolved window policy changing. The reason is that UsdImagingDelegate provides a window policy for cameras (even though window policy is not a concept for a UsdCamera) which by default is CameraUtilMatchVertically. The UsdImagingStageSceneIndex does not provide a window policy (and shouldn't since it is not a USD concept). Without this change, using the scene index instead of the delegate means that some tests switch the window policy. With this change, the test explicitly demands a window policy which can be configured from the command line.

(Internal change: 2271219)