Add Support for filtering processing of assetInfo.identifier metadata
This change adds in functionality for the FileAnalyzer to skip processing of assets declared in specified metadata keys.  The motivation for this change was from Github Issue #2135.  In this case the assetInfo.identifier was present in a flattened usda file that was being usdz'd.  This caused the original usda file to be present in the archive as well.

This change specifically ignores processinfg assetInfo.identifier.  Future work would invove expanding the keys that are not processed or allowing the user to specify such keys.  Currently metadata filtering is only enabled for creating usdz files.

Note: although the key is not procesed, its value is still remapped.  This prevents leaking user file system paths into the archive when creating arkit flavored usdz files or zipping a flattened file as in the original github issue.

Fixes #2135

(Internal change: 2271349)