Add support for the ri:integrator in usdview.
The Integrator follows the same pattern as the Sample and Display
Filters, in that they gather the parameters and shaderId into a
HdMaterialNode2 which is then used to create the correcponding
riley node needed for the integrator. Again same as with the filters,
when the RenderSettings prim is synced it will let
HdPrman_RenderParams know the path of the connected integrator,
so when the integrator is synced, it will know to create the riley
node if it is the connected integrator.

The test harness is also updated to use a fallback Integrator with the RenderSettings Prim.

(Internal change: 2271443)