build_usd.py will now error out if run with Python 2.
This is a follow-up to the removal of Python 2 build support
in change 2269075. Without this change, users running the
build script with Python 2 would only hit an error when they
reached the USD build step after building dependencies.

This is a minimal change to block users from running the script
with Python 2. A further follow-up will remove the remaining
code related to Python 2.

(Internal change: 2272064)