tweak for MatX on by default for latest Clang
There was an extraneous semicolon in the script.
MaterialX 1.38.4 uses an outdated Google Test that uses deprecated parts of std, which have been removed in the latest iteration of Clang as shipped with current XCode. This tweak disables building the MaterialX test suite in order to avoid that issue. v1.38.7 or 8 will have a fix, although arguably we shouldn't be spending the time to build the MaterialX test suite here, as part of the OSS build_usd.py regimen.

(Internal change: 2272166)