[HdSt] Reorganized Metal PTVS meshShaderKey source snippets
This cleans up the set of shader source snippets used to manage face
and edge parameterization for PTVS shaders to simplify some control
flow in meshShaderKey and to have fewer or when necessary more
clearly named platform specific snippets.

Plumbing of PTVS position in patch coordinates has been moved to
codeGen so that we don't have to deal with this separately in each
PTVS shader. This interstage value has been renamed to "hd_tessCoord"
instead of "ptvsBarycentricCoord".

Triangulation of quad faces in the GS now emits sub-triangle
vertices in the same order as tri-quad triangulation, i.e.
v0,v1,v2 and v2,v3,v0. This is a step along the way to eliminating
the remaining use of GS stages.

(Internal change: 2272357)