Exclude pegtl from doxygen input files
The single pegtl header causes doxygen (v1.8.14, at least) to
seemingly run forever. From a stack trace, doxygen appears to spend
all its time in a function called "findInheritedTemplateInstances".
My guess is that the heavy template usage in the header is too much
for doxygen to handle.

This is a third-party library and does not contain any doxygen
tags anyway, so we lose nothing by excluding it.

(Internal change: 2273080)