Fix export of USD as a monolithic shared library.
Outside projects could not easily be built against USD as a monolithic
shared library because of some usage requirements not set on the
exported usd_ms target. For example, the INTERFACE_INCLUDE_DIRECTORIES
target property was missing.

This is fixed by reverting the use of the $<BUILD_INTERFACE:...> export
generator expression.

However, outside builds would still fail because the usd_m and other
internal-to-monolithic targets were not exported. This is fixed too.