[hdSt, hgi*] Guard against correctness issues resulting from GPU instance frustum culling of prims with geom subsets.
Storm can create draw items for each of the geom subsets of a prim; these geom
subset draw items will share the same instanceIndices and culledInstanceIndices
buffers. This can result in a correctness issue during GPU instance frustum
culling, which will have a separate shader invocation per geom subset per
instance, as different threads will be attempting to write to the same
culledInstanceIndices buffer with potentially different values, depending on
which threads execute the operation first.

Thus we add an extra value to the start of the instanceIndices and
culledInstanceIndices buffers. The instance frustum cull shader performs an
atomic operation to update this value to the value of the xor of the thread's
drawCommandIndex and drawBatchId. Since only one thread per shared
culledInstanceIndices buffer will be successful, we can then use that value to
allow only one geom subset draw item to write to the shared
culledInstanceIndices.

Fortunately, this is not an issue for the compute version of GPU instance
frustum culling used by the pipeline draw batch. If we ever delete the vertex
GPU instance frustum cull shader, we might be able to revert this change.

Fixes #1892

(Internal change: 2273125)