[Hd,HdSt] Moved Storm specific aspects of smooth/flat normals
The computation aspects of the smoothNormals and flatNormals
implementation are used only by Storm. The CPU specific versions
of these aren't currently in use, but they remain a useful reference
so just relocating the code for now, not deleting it yet.

The remaining aspects in Hd might eventually be moved elsewhere,
e.g. geomUtil, but not doing so as part of this change.

(Internal change: 2273603)