[Hd,HdSt] Moved Storm computation from hd to hdSt
This is used only by Storm. It is still named as a public
class since it is exposed in other public classes, but further
refactoring may allow this to eventually become a true private
class within the Storm implementation.

The updated HdStComputationComputeQueuePairVector type alias is
awkward and we may revisit this naming as part of adding support
for more explict queueing of chained dependent computations.

(Internal change: 2274375)