Adding HdDataSourceLocatorSet::insert(&&) with some optimizations.
UsdImagingStageSceneIndex::_InvalidateImagingSubprim was essentially just trying to compute the union of several HdDataSourceLocatorSet using HdDataSourceLocatorSet::insert.
I noticed that it had an optimization avoiding the call to insert for the case of having only one set. And that during the refactor, the call to _InvalidateImagingSubprim was dropped.

I figured that it is better to achieve a similar optimization but at a lower level, namely, HdDataSourceLocatorSet itself.

(Internal change: 2274767)