pcp: Remove PcpLayerStackIdentifierStr and PcpSiteStr
These objects were introduced to allow PcpErrorBase
and subclasses to store the source location of an error
in a stable form independent of the lifetime of layers
and other objects. This would theoretically allow clients
to hold on to and query error objects beyond the lifetime
of the associated prim index or layer stack.

In practice, these error objects are just used by clients
immediately after prim index computation and not retained
afterwards. USD, for example, just emits warning messages
for each errors immediately. Also, many error objects were
not using PcpLayerStackIdentifierStr/PcpSiteStr and were
holding on to layer handles instead. This has been the case
for a long time, and no issues related to this have been
reported.

Given the limited use and deficiencies noted above, this
change just removes PcpLayerStackIdentifierStr and PcpSiteStr.
The Pcp error objects now just hold on to layer handles
instead. This allows us to avoid a non-trivial update to
PcpLayerStackIdentifierStr which would be needed due to
an upcoming change that will add at least one new field
to PcpLayerStackIdentifier.

If we do need a persistent representation for Pcp errors
in the future, I think we'd need a bigger redesign of the
error objects to allow Pcp to invalidate them when
associated prim indexes, etc. are invalidated.

(Internal change: 2274951)