Adding UsdImagingStageSceneIndex support for showing unloaded prims as bounding boxes.
To activate this mode, the scene index needs to be created with a flag to include unloaded prims and a filtering scene index added that switches the unloaded prims to draw mode bounds.

The stage scene index needed to populate metadata to indicate whether a prim is loaded.

The draw mode scene index had to also be able to consume the extentsHint when there are no authored extents.

(Internal change: 2275832)