UsdImaging_AdapterManager refactor:
Replacing std::pair APISchemaEntry by struct AdapterEntry with comments explaining the members.
Making main function on UsdImaging_AdapterManager return a const reference to an AdaptersEntry struct (which was private) to avoid copying an array and avoid having a different mechanism for returning all adapters vs just the prim adapter.
Renamimg internal types and variables to be more consistent.
Factoring out the code that is computing all adapters from the code that is caching the result.

(Internal change: 2276039)