[hd, hdsi, usdImaging] Use closure registration for the scene globals scene index (SGSI)
Motivation:
SGSI differs from scene indices that use plugin-based registration in that it is an application-facing scene index that reflects state that either isn't backed by scene description, or is driven by the application. The change that introduced SGSI had the application insert it into the merging scene index. This adds overhead to each GetPrim call because it is added at the root path, even though it is responsible for just the sceneGlobals locator at that path. Using the recently added closure-registration allows the client to retain a handle to the scene index (and thus configure it as necessary) while using a similar treatment to scene index plugins. This avoids adding it directly to the merging scene index.

- Update SGSI to be a filtering scene index. Validate the path returned when querying the data source instead of in the helper utility.
- Add templated facility HdUtils::RenderInstanceTracker to associate application objects managed by a shared_ptr with a renderInstanceId.
- Update UsdImagingGLEngine to:
-- register a callback to create and append the SGSI.
-- generate a unique render instance id for use in render index construction and for associating application scene indices.

(Internal change: 2276520)