[usdImaging] Nested point instancers aren't correctly transformed
When a point instancer is within the prototype of another instancer, it may be incorrectly transformed when a transform is authored on the prototype root. UsdImagingPointInstancerDelegate only captured transforms up to the prototype root, but not on the prototype root itself. This is inconsistent with the treatment of other, non-point-instancer prims within the prototype subtree, which do correctly take the transform on the prototype root.

This fix corrects this behavior by causing UsdImagingPointInstancerAdapter to return the point instancer's transform relative to the prototype root's parent, rather than relative to the prototype root. This approach will produce the correct relative transform regardless of the parent instancer type (point vs native), or of whether the prototype root is a descendant of the parent instancer.

An additional test case is also added to testUsdImagingGLPointInstancer to cover this issue.

Fixes #2359

(Internal change: 2277872)