[usdImaging] Fix for transforms on native-nested point instancers
Legacy UsdImaging uses adapter hijacking to handle native instancing. UsdImagingDelegate associates prims within a native instancing prototype with the native instancing adapter, rather than the proper adapter for their type. The native instancing adapter maintains the link to the proper adapter. This means that all calls on the native instancing adapter should check whether the supplied cachepath is in an instancer's prototype and, if so, delegate the call to the proper adapter for the prototype prim type. In the identified problem case, the prototype prim is a point instancer.

UsdImagingInstanceAdapter::GetInstancerTransform did not make this check. It assumed it would only ever be called for the transform of a native instancer and simply returned the root transform, since all native instancers are relocated to root.

The fix is simple: GetInstancerTransform needs to check whether the instancer whose transform it's been asked for is actually a prototype prim within a native instancing prototype. When it is, it should instead forward the call to the prototype prim's own adapter.

Fixes #1663

(Internal change: 2279634)