UsdImagingStageSceneIndex: Support for coordinate system in hdPrman.
Note that a coordinate system binding in USD can target any xformable prim, but hdPrman requires coord system prims.

Thus, we need to create coordSys hydra prims under the targeted xformable prim with the same xform for hdPrman. UsdImagingDelegate did this in the scene index. We do this through a scene index plugin.

Removing code that attempted to create coordSys prims in the prim adapter.

The HdCoordSys prim also got that the name for the coord system from a particular naming convention. This change makes HdCoordSys query the name from the scene delegate which in emulation corresponds to the name in the HdCoordSysSchema.

(Internal change: 2281744)
(Internal change: 2281754)
(Internal change: 2282041)