Fix for PySide 6.5.0 initialization order.
The recent change in Pyside 6.5.0 here:

https://github.com/pyside/pyside-setup/commit/e8095467f7d0332cc0987e7c541de9906e19fece

Introduced correct multiple inheritance through the use of the `super()` Method Resolution Order.  I checked the MRO of ViewSettingsDataModel, which was:

```
<class 'pxr.Usdviewq.viewSettingsDataModel.ViewSettingsDataModel'>
<class 'PySide6.QtCore.QObject'>
<class 'Shiboken.Object'>
<class 'pxr.Usdviewq.settings.StateSource'>
<class 'object'>
```

This meant that the `StateSource` object wasn't being initialised correctly.  With this change the MRO is:

```
<class 'pxr.Usdviewq.viewSettingsDataModel.ViewSettingsDataModel'>
<class 'pxr.Usdviewq.settings.StateSource'>
<class 'PySide6.QtCore.QObject'>
<class 'Shiboken.Object'>
<class 'object'>
```

And the initialisation works correctly.  Checked with PySide 6.5.0 and 6.4.3

Contribution: Apple

Fixes #2392

(Internal change: 2281902)