Rolling back a code review note that introduced a Hydra notice behavior regression
Flattening scene index should not return a dataSource
for prims that do not exist in the input scene, because
it can confuse downstream scene indexes as to the existence
or non-existence of prims.

In particular, the Merging scene index is sensitive to this,
since it will convert Removes to Adds if one of its inputs
reports a valid datasource.  This was therefore always the case
if one of the inputs was a Flattening scene index.

Reviewing the code review notes, the theory there was that
we might need to support non-contiguous namespaces and flatten
attributes down across them, such as /World and /World/scope/A,
without "scope" in the middle existing.  However, that does
not seem to be the case in current practice / our test suite,
and accomodation of that theory introduced this problem, so
let's revisit the theory in the future should it occur.

(Internal change: 2282264)