Introduce OpenEXRCore to Hio
This change is a prerequisite for adding built-in OpenEXR image reading for Hio.

This change includes the sources for the OpenEXRCore 3.2.0 library, and a unity build file that compiles the sources into the Hio library. The sources are concatenated into a unity build, in order that all symbols are available within a single compilation unit. All functions are adorned with either an API macro, or an INTERNAL macro, both of which are set to static to prevent symbols from leaking out of Hio. In this way, the interred OpenEXRCore library will not conflict with any user-preferred OpenEXR libraries that may be on the system.

A subsequent change will introduce the EXR reading plugin itself, however this change does not include it, in order to manage the complexity and review burden of interring the initial OpenEXRCore library.

Note that some includes have been modified to use quoted paths instead of angle paths, and certain symbols have had an extern marking removed in order to prevent leaking those symbols. All such modifications are marked with "pxr" in a comment. There are very few. The INTERNAL adornments are still in progress to be merged upstream, but are not marked "pxr" as there shouldn't be an issue in landing them upstream.

All other modifications to the OpenEXRCore have already been landed upstream.

With regards to the maintainability of the unity concatenation, the OpenEXR team has indicated a desire to have the concatenated build available as part of the standard code distribution. As such, any maintenance with respect to that file will be handled by CI on the EXR project itself. A unit test will be landed upstream to ensure this is reliably in place, however the exact approach is still in discussion.

(Internal change: 2282579)
(Internal change: 2282745)