usd: Add support for stage variable expressions
UsdStage will now evaluate any stage variable expressions in
asset path-valued attributes and metadata. The evaluated
result will then be anchored and resolved just like a
regular asset path would.

When expression variables are changed in a layer stack,
UsdStage will send a UsdNotice::ObjectsChanged notice with
resyncs for any prims that depend on that layer stack. In
particular, this means that a resync will be sent for the
entire stage in the common case where variables are modified
in that stage's root layer.

This behavior reflects the fact that any asset-valued
attributes and metadata on those prims or their descendants
may be affected by these variable changes. UsdStage is
currently unable to send more targeted notifications, since
it does not cache any of these values and does not want to
incur the cost of scanning the entire stage (and potentially
pulling lots of data into memory) to discover them. Note,
however, that the UsdStage itself does not need to recompose
any prims when variables are changed, unless those variables
are involved in composition arcs.

We plan to replace this scheme with a new asset path
invalidation notice as one of the first improvements to
this feature after its release. In the meantime, the current
over-invalidation behavior ensures that users will see
the correct results when expression variables are changed
(thought at potentially high performance costs) and
allow us to roll out this feature for initial testing.

(Internal change: 2282631)