Make GfBbox3d trivially copyable by letting the compiler provide a default copy constructor.
This initially helped address GCC9 issues related to libraries that used
this type in a way that was unsafe when it wasn't trivially copyable.
The compiler errors will be addressed by using the correct test for
whether types are trivially copyable, but this change will allow
GfBbox3d to take the fast path in that code.

This change also allows a default move constuctor (and copy and
move assignment operators) to be generated.

(Internal change: 2282855)