Fix crash in certain cases where "use scene materials" is disabled.
Change 2277507 introduced a new approach to toggling usage of scene
materials based on a scene index that filters out material prims
and material bindings.

In this crash, a Storm task was failing to find a Storm render
buffer corresponding to the 'color' AOV binding. The reason is that
the HdsiPrimTypePruningSceneIndex was configured to prune scene
materials.  This pruning removed not only preview materials, but
also child prims in the scene index and associated renderBuffers.
However, tasks that depended on these render buffers were being
added directly into the render index. Consequently, we are left
with a task (and render pass) in Storm, but not associated
render buffers to bind.

This change modifies the pruning scene index to simply
null out the primType and dataSource of pruned prims,
without pruning the entire subtree beneath them.  This
allows the renderBuffers, etc. to pass through.

(Internal change: 2283392)