[hdPrman] PbsPathTracer should render volume aggregates by default
We will enable volume aggregate rendering when there is no authored integrator node in the scene and the legacy render settings map (or HD_PRMAN_INTEGRATOR) directs us to use PbsPathTracer.

This behavior is different from the normal behavior of PbsPathTracer in Renderman, which is to disable volume aggregate rendering by default. Therefore, we limit our override of this default to situations where there is no authored integrator node in the scene.

This change also reverts 2279759.

If there is a valid, active PbsPathTracer node, we make no changes to it, even when it is silent on the subject of volume aggregate rendering. If the scene uses an authored PbsPathTracer node and volume aggregates should be rendered, the PbsPathTracer node must say so.

(Internal change: 2284355)