usd: Fix (maybe) intermittent failures in testUsdInstancingCpp
I believe the problem is that the test cases would store pointers
to prototype source prim indexes before and after a scene description
change and check if they differed to indicate whether that prim index
had been recomputed by Pcp. This worked most of the time but could
fail if the recomputed prim index happened to be placed at the same
memory address as the old one.

I was unable to confirm this theory locally but I'm modifying the
test defensively to avoid this issue. Rather than checking whether
a prototype's prim index has been recomputed, the test now checks
whether a prototype's source prim index path has changed. This is
more pertinent to what this test is trying to verify anyway.

(Internal change: 2284399)