[hd][hdSt] Hulls should respect hull opacity
The goal was to achieve behavior similar to that of wireframe + surface mode while preserving the ability to control the hull color with a mesh's hullColor parameter. The hullColor parameter has long controlled the *color* of hull wires, but its *opacity* channel has been ignored in storm. This change makes storm respect the hullColor opacity channel.

To achieve this without changing the appearance of hulls elsewhere, we add a parameter to HdMeshReprDesc by which applications may specify whether hulls and wireframes (collectively "edges") should always be fully opaque. With this added parameter, we are able to preserve existing hull and wireframe appearance in UsdView. Note that this parameter applies to hulls ("hull edges") and wireframes ("edges"), depending on the selected MeshGeomStyle.

hull opacity is controlled by the alpha channel of the mesh's hullColor parameter. The default value for hullColor.a has not been changed, but since it is 0.7 users will likely see the change immediately.

(Internal change: 2285298)