Make RenderColorWithOccludedSelection take non-opaque colors into account.
Any non-opaque geometry which would normally be rendered via alpha-to-coverage
cannot do so when using RenderColorWithOccludedSelection since it requires
blending.  To deal with this, we instead use the standard blending approach for
non-opaque geometry, and if there is an occluded selection behind this geometry
we also increase the contribution from the selected geometry when blending.

(Internal change: 2287071)