hio: use a named struct for stbi__png typedef in stb_image.h and update patch
The local patch to stb_image.h adds a "gamma" member to the previously unnamed
struct used in the typedef of stbi__png. It also includes a default initializer
to set the member to 0.

When using C++17 or later, some compilers will emit an error that an unnamed
class cannot be used within a typedef if it has a non-static data member with a
default member initializer. In particular, this is the case with msvc:

https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/c5208

In this case, the error can be avoided simply by giving the problematic struct
a name. stb_image.h in hio was updated to make this change, and the patch file
used to create it from the original source version was regenerated to ensure
that the same change is applied if/when stb is upgraded in the future.