Fix uniform name conflicts in Storm
Storm injects a built-in scale and bias uniform for texture types. When using materialx graphs, this can cause name collision when materialx inputs.
For example if a node has inputs called normal (texture) and normal_scale (float), this will conflict when storm adds a normal_scale (for texture).
The workaround or fix is to use a unique name. We use a named suffix _hdSt_ to define Storm generated uniforms.