[build] Disable warning C4003 on Visual Studio builds
Changes 2287201 and 2290829 removed the use of
tf/preprocessorUtils.h in several header files, which caused
thousands of "not enough actual parameters for macro" warnings
in the build. This was because preprocessorUtils.h defined
pragmas and macros that disabled this warning, which
propagated throughout the build via header includes.

Rather than reapplying the macros to the affected headers,
this change just disables the warning globally. This
centralizes the behavior and makes it easy for us to
turn it off if/when we move to the "new" standards-compliant
Visual Studio preprocessor.

Since the macros disabling the warning are no longer in headers,
client code building against USD may now start to see these
warnings and will need to disable the warning in their own
builds. This is arguably better, as having USD affect external
build behavior is not very neighborly.

See comments in the change for more details.

(Internal change: 2292363)