Version up DomeLight to DomeLight_1, which introduces the "poleAxis" property. This property affects the dome's final rendered orientation and allows for orientation-dependent properties like texture maps to retain the user's desired orientation when the stage's coordinate system changes.
For UsdImaging, the DomeLight_1's new adapter uses the "poleAxis"
property to create and inject a "domeOffset" matrix into the subprim
data and/or supply it when directly requested. hdSt and hdPrman look for
the "domeOffset" when creating/updating their light objects.

Updated tests:
- testUsdLuxLight: Now that we have 2 dome light types that correspond
to the same Sdr "node", it makes sense to break out the expected types
and expected nodes into 2 different lists.

(Internal change: 2292895)
(Internal change: 2292922)