Update HdPrmanRenderPass to use RenderSettings prim
HdPrman_RenderPass will now update the cameraContext based on
the render settings prim, if we are rendering based on the render
settings prim, it is a valid render settings prim, and we are not in
rendering interactively.

A render settings prim is considered valid if there is at least one
render product and there is a camera relationship. Note that in
the above scenario we are not using any aov bindings, so they
need to all be marked as converged, to avoid waiting indefinitely
for them to converge.

(Internal change: 2293235)