Update UsdRecord to use the RenderSettings prim
When HD_PRMAN_RENDER_PASS_USE_RENDER_SETTINGS_PRIM
is set to true usdrecord will generate image(s) based on the
RenderProducts authored on the RenderSettings Prim.

When the environment variable is set to true, if there is a
'renderSettingsPrimPath' listed in the stage metatdata the resulting
images will be generated from the authored RenderProducts. The
renderSettings prim can also be indicated by specifying the path with
the '--renderSettingsPrimPath' argument, if you want to override the
stage metadata.
Note that authored attributes on the RenderSettings/Product prims will
override the imageWidth and camera usdrecord arguments, and ignore
the outputImagePath argument.

(Internal change: 2293236)