usd: Add convenience c'tor for UsdNotice::ObjectsChanged
Add a separate c'tor for the ObjectsChanged notice that only
takes a map of resync changes. This simplifies callsites that
were previously forced to instantiate empty maps for info
changes themselves, and will make it easier to add additional
change maps in an upcoming change.

As an alternative, I could have just allowed client code to
pass in nullptr for the info change map. I decided against
this because I'd have to add checks for nullptr throughout
the ObjectsChanged notice implementation, with all the
risks and costs that entails.

(Internal change: 2293243)