pcp: Fix missing resyncs when handling asset resolver changes
Asset resolver changes affecting sublayer asset paths in a
layer stack would not register resyncs for prim indexes
using that layer stack.

As it happens, the change processing needed for asset resolver
changes is the same as what's needed when a layer's resolved
path changes -- which makes sense since an asset resolver change
can affect a layer's resolved path. So we can just reuse the
existing code to handle the latter case.

This fix will be exercised in a unit test in an upcoming
change.

(Internal change: 2293305)