usd: add a UsdSchemaRegistry function for identifying schematics layers
This adds a UsdSchemaRegistry::IsSchematicsLayer() function that can be used to
determine whether a given layer is a schematics layer stored in the schema
registry. Getting the list of all loaded layers can be useful when debugging
lifetime issues, and this new function makes it easier to filter that list to
exclude schematics layers since they will persist indefinitely.