usdAbc: Add ImathConfig to target link libraries
At least one Alembic header includes half.h as <half.h>,
which is not the canonical include path (<Imath/half.h>)
when building against Imath 3. This requires adding the
"include/Imath" directory to the include paths when building
usdAbc.

For some yet-unknown reason, this was not happening
specifically in monolithic builds, which caused builds to
fail due to not finding the half.h header. This was observed
when building USD against Alembic 1.8.5 and OpenEXR 3.1.11.

This workaround adds the Imath::ImathConfig target to usdAbc
to go along with Imath::Imath. This puts the "include/Imath"
directory back into the include paths, avoiding the issue.

Note that other libraries like hioOIIO and hioOpenVDB also
use the Imath::Imath. However, OpenImageIO and OpenVDB have
built-in handling for the different half.h header locations
so we don't need to do anything for them.

(Internal change: 2295815)